# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 명지피앤피 인사·총무 DX 시스템 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedules/iso:
    get:
      operationId: listIsoSchedules
      tags: [schedules]
      summary: ISO 일정 목록
      responses:
        "200":
          description: ISO schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IsoSchedule"
    post:
      operationId: createIsoSchedule
      tags: [schedules]
      summary: ISO 일정 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IsoScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IsoSchedule"
  /schedules/insurance:
    get:
      operationId: listInsuranceSchedules
      tags: [schedules]
      summary: 보험 갱신 일정
      responses:
        "200":
          description: Insurance schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InsuranceSchedule"
    post:
      operationId: createInsuranceSchedule
      tags: [schedules]
      summary: 보험 갱신 일정 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsuranceScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsuranceSchedule"
...
components:
  schemas:
...
    IsoSchedule:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        scheduledDate:
          type: string
        type:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - scheduledDate
        - type
        - createdAt
    IsoScheduleInput:
      type: object
      properties:
        title:
          type: string
        scheduledDate:
          type: string
        type:
          type: string
        description:
          type: string
      required:
        - title
        - scheduledDate
        - type
    InsuranceSchedule:
      type: object
      properties:
        id:
          type: integer
        insuranceName:
          type: string
        renewalDate:
          type: string
        insurer:
          type: ["string", "null"]
        amount:
          type: ["integer", "null"]
        daysUntilRenewal:
          type: integer
        urgency:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - insuranceName
        - renewalDate
        - daysUntilRenewal
        - urgency
        - createdAt
    InsuranceScheduleInput:
      type: object
      properties:
        insuranceName:
          type: string
        renewalDate:
          type: string
        insurer:
          type: string
        amount:
          type: integer
        notes:
          type: string
      required:
        - insuranceName
        - renewalDate
...