# lib/api-zod/src/generated/types/insuranceSchedule.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * 명지피앤피 인사·총무 DX 시스템 API
 * OpenAPI spec version: 0.1.0
 */

export interface InsuranceSchedule {
  id: number;
  insuranceName: string;
  renewalDate: string;
  /** @nullable */
  insurer?: string | null;
  /** @nullable */
  amount?: number | null;
  daysUntilRenewal: number;
  urgency: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}