# lib/api-zod/src/generated/types/insuranceScheduleInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * 명지피앤피 인사·총무 DX 시스템 API
 * OpenAPI spec version: 0.1.0
 */

export interface InsuranceScheduleInput {
  insuranceName: string;
  renewalDate: string;
  insurer?: string;
  amount?: number;
  notes?: string;
}